Fix broken build via 'gulp build'. Remove console.log in testcases.
Broken build process caused by es6 code is used especially => (arrow for
function declaration). Thus we decided to use babel to convert it to
es2015 first before continue building process.

We don't use uglify-js-harmony at the moment as it's compicate to
integrate with our current build task in gulpfile.js.

This means we need to use
- minifier from gulp-uglify
- uglify from uglify-js

When minifier and uglify from gulp support es6, we will update this
things again.