Add handling no Internet connection when making a request for http/https
The error message that thrown from calling SDK regarding to the case of
no internet connection depends on which browser SDK is running on.

This change added a private utility function to check of those message
as well and return boolean whether it should add error code as handled
in application level by developers or not.

Bumped to version 1.3.7.