Add auto token request for POST request (if needed)
Auto token request will be carried out if such a current POST request
which always needs valid token to be present has no token (null), or
expired token.

It checks against the error code of 900 and 902. If so then it will make
duplicate effect of request of Playbasis.authApi.auth() but coded in
low-level way in http.js. This is to avoid cyclic dependency.

Test cases are also added in auto.req.token.test.js and tests against
both case of need and no-need for such token request.