Fix #1. Update karma fast to test against Chrome.
The root problem is due to the code only has listened to 'data' event
and treated it as a final response from end-point; it's not the case.

'data' event fired when there's a new chunk of data ready for client to
consume. 'end' event is another important event that needs to be
listened to too. The latter will serve as a final event to combine all
chunk of data together and process our normal routine on received data.

We use normal array to combine all string data for each chunk, then
combine it via .join('') and not even use Buffer.concat() or anything
that if we use it we might have to add another dependency on NPM as
Buffer is not natively supported on browser.

Also test case against requesting for GET method is added and use engine
api's listRules for testing application that we have plenty of data to
return.

For POST method, we didn't have such end-point that can return that
large amount of data yet, so it won't be included. But the code is
implemented in the same way to prevent the error like in GET method.

Note
Testing code via NodeJS against different browser will give different
result about the number of times 'data' event would be fired. This is
probably due to buffer length as set differently on each browser. Thus
when testing keep an eye on which browser you test against too to avoid
wasting time that no multiple of 'data' event fired.

For data size around ~280k bytes of end-point can return when request
- Firefox -> generates only 1 time 'data' event
- Chrome -> generates 3 times of 'data'` event